split utils into utils and unzip
