+++ lib/node/index.js
@@ -12,7 +12,8 @@ var format = require('url').format;
 var resolve = require('url').resolve;
 var methods = require('methods');
 var Stream = require('stream');
-var utils = require('./utils');
+var utils = require('../utils');
+var unzip = require('./unzip').unzip;
 var extend = require('extend');
 var mime = require('mime');
 var https = require('https');
@@ -787,7 +788,7 @@ Request.prototype.end = function(fn){
 
     // zlib support
     if (self._shouldUnzip(res)) {
-      utils.unzip(req, res);
+      unzip(req, res);
     }
 
     if (!parser) {
