+++ lib/node/response.js
@@ -4,7 +4,7 @@
  */
 
 var util = require('util');
-var utils = require('./utils');
+var utils = require('../utils');
 var Stream = require('stream');
 
 /**
